Code moving and add reentrant locks
